feat: move cairo 2 events to single indexer
