+++ env
@@ -4,6 +4,7 @@ AR_FINALITY='DATA_STATUS_ACCEPTED'
 STARTING_BLOCK=911736
 AR_STARTING_BLOCK=911736
 ID_UPGRADE_A_BLOCK=911736
+NAMING_UPGRADE_A_BLOCK=915000
 MONGO_CONNECTION_STRING='mongodb+srv://login:BTiTxKFdfkU5H54h@testing.rzwua3b.mongodb.net/?retryWrites=true&w=majority'
 ETH_CONTRACT='0x049D36570D4e46f48e99674bd3fcc84644DdD6b96F7C741B1562B82f9e004dC7'
 IDENTITY_CONTRACT='0x29a1b22ae04ec2d06fd15608ab1585e334fcb3d501b908c78434cf2ee6435a0'
